% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCR.R
\name{CRlearning}
\alias{CRlearning}
\title{CR Learning}
\usage{
CRlearning(x, ...)
}
\arguments{
\item{x}{An object used to select a method}

\item{...}{Additional arugments passed to other methods.}
}
\description{
Strengthening of Target feature --> Target feature associations
}

