% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall.R
\name{cuedRecall}
\alias{cuedRecall}
\title{PCR Cued Recall Simulation}
\usage{
cuedRecall(x, cue = 1, ...)
}
\arguments{
\item{x}{An object used to select a method. The PCR package implements methods
for \code{timed} and \code{PCR} objects.}

\item{cue}{A scalar numeric indicating which cue to test on}

\item{...}{Additional arugments passed to other methods.}
}
\description{
PCR Cued Recall Simulation
}

