% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxicausal.R
\name{hbridge}
\alias{hbridge}
\title{This function is to compute estimating equation of outcome-inducing confounding bridge function}
\usage{
hbridge(para, Y, W, Z)
}
\value{
returns the sample level estimating equations for q function
}
\description{
This function is to compute estimating equation of outcome-inducing confounding bridge function
}
\keyword{internal}
