% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCR.R
\name{initPCR}
\alias{initPCR}
\title{Title}
\usage{
initPCR(params, distribution = "beta", time = 0, nFeatures = 1000,
  nSims = 1000, nItems)
}
\arguments{
\item{params}{A named list of numerics to be used as PCR model parameters.}

\item{distribution}{A character vector naming the underlying distribution of
feature activation and thresholds. Can be either \code{"beta"} or \code{"binomial"}.}

\item{time}{A numeric scalar indicating the duration of the trial. If set to 0, then
retrieval latency is not simulated by PCR functions.}

\item{nFeatures}{A scalar numeric value giving the number of features each item
is assumed to have in memory.}

\item{nSims}{A scalar numeric value giving the number of times a memory process
should be simulated in the model.}

\item{nItems}{A scalar numeric value giving the number of items to be remembered.}
}
\value{
A PCRparams object
}
\description{
Title
}
\examples{
param_list <- list(ER=.6, LR=.15, FR=.1, TR=.1, TV = .05,
                   Tmax = 75, Tmin = 0.5, lambda = .5)
beta_timed <- initPCR(params = param_list, distribution = "beta", time = 10,
                      nFeatures = 100, nSims = 1000, nItems = 20)
str(beta_timed, give.attr=FALSE)
}

