% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCR.R
\name{initPCRparams}
\alias{initPCRparams}
\title{Create a PCRparams object to control be used by PCR functions like \code{\link{study}}.}
\usage{
initPCRparams(params, distribution = "beta", time = 0, nFeatures = 1000,
  nSims = 1000, nItems)
}
\arguments{
\item{params}{A named list of numerics to be used as PCR model parameters.}

\item{distribution}{A character vector naming the underlying distribution of
feature activation and thresholds. Can be either \code{"beta"} or \code{"binomial"}.}

\item{time}{A numeric scalar indicating the duration of the trial. If set to 0, then
retrieval latency is not simulated by PCR functions.}

\item{nFeatures}{A scalar numeric value giving the number of features each item
is assumed to have in memory.}

\item{nItems}{A scalar numeric value giving the number of items to be remembered.}

\item{nSim}{A scalar numeric value giving the number of times a memory process
should be simulated in the model.}
}
\description{
Create a PCRparams object to control be used by PCR functions like \code{\link{study}}.
}
\examples{
a <- initPCRparams(params = list(ER=.6, LR=.2, FR=.1, TR=.15, TV = .05),
                   distribution = "beta", nItems = 20, nSims = 1000,
                   nFeatures = 100, time = 10)
}

