% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxicausal.R
\name{MSE_func}
\alias{MSE_func}
\title{This function is to compute mean squared error}
\usage{
MSE_func(bridge_func, para, Y, W, Z)
}
\value{
returns mean squared error
}
\description{
This function is to compute mean squared error
}
\keyword{internal}
