% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCR.R
\name{PRlearning}
\alias{PRlearning}
\title{PR Learning}
\usage{
PRlearning(x, ...)
}
\arguments{
\item{x}{An object used to select a method}

\item{...}{Additional arugments passed to other methods.}
}
\description{
Strengthening of Cue --> Target associations on un-encoded features
}

