% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxicausal.R
\name{qbridge}
\alias{qbridge}
\title{This function is to compute estimating equation of treatment-inducing confounding bridge function}
\usage{
qbridge(para, Y, W, Z)
}
\value{
returns the sample level estimating equations for q function
}
\description{
This function is to compute estimating equation of treatment-inducing confounding bridge function
}
\keyword{internal}
