% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6.R
\name{reaction_times}
\alias{reaction_times}
\title{Extract observed reaction times}
\usage{
reaction_times(object, ...)
}
\arguments{
\item{object}{An R object}

\item{...}{Arguments passed to other methods}
}
\description{
Extract observed reaction times
}

