% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studying.R
\name{study}
\alias{PRlearning}
\alias{study}
\title{Study}
\usage{
study(x, cue = 1, ...)
}
\arguments{
\item{x}{An object used to select a method}

\item{...}{Additional arugments passed to other methods.}
}
\description{
Effects of studying on un-encoded features
}

