% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCR.R
\name{summary.PCR}
\alias{summary.PCR}
\title{Summarize a PCR model condition}
\usage{
\method{summary}{PCR}(x)
}
\arguments{
\item{x}{A PCR model object}
}
\value{
A data frame
}
\description{
Summarize a PCR model condition
}
\examples{
param_list <- list(ER=.6, LR=.15, FR=.1, TR=.1, TV = .05,
                   Tmax = 75, Tmin = 0.5, lambda = .5)
beta_timed <- initPCR(params = param_list, distribution = "beta", nItems = 20,
                            nSims = 1000, nFeatures = 100, time = 10)

init_study <- summary(beta_timed)
restudy <- summary(init_study)
summary(init_study)
summary(restudy)
restudy <- summary(beta_t_restudied)
}

