% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/memories.R
\name{test}
\alias{test}
\title{test}
\usage{
test(mem, nFeatures, thresh, acc, LR, TR, FR = NULL)
}
\arguments{
\item{mem}{Matrix of memory strengths.}

\item{nFeatures}{Number of features the memory for each item has}

\item{thresh}{Matrix of memory thresholds. Threshold is the number of active features
needed to support pattern completetion}

\item{acc}{Logical matrix of performance for previous test}

\item{LR}{Learning Rate. Gives the probability of learning a new features
through practice}

\item{TR}{Threshold Reduction rate. Describes the probability of removing the need
 for a remembering a particular feature following successfull recall}

\item{FR}{Forgetting Rate. Gives the probability of forgetting a feature}
}
\value{
List of two matrices. mem is a matrix of memory strengths following succcesful
 recall. thresh is a mastrix of feature thresholds following succesful recall.
}
\description{
test
}

