% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/recall.R
\name{cuedRecall}
\alias{cuedRecall}
\title{PCL Cued Recall Simulation}
\usage{
cuedRecall(mem, thresh, space = NULL, Tmin = NULL, Tmax = NULL,
  Time = NULL, lambda = NULL)
}
\arguments{
\item{mem}{}

\item{thresh}{}

\item{space}{}

\item{Tmin}{}

\item{Tmax}{}

\item{Time}{}

\item{lambda}{}
}
\value{
Logical Matrix representing recall success/failure of each list item
 for each simulation
}
\description{
PCL Cued Recall Simulation
}

