% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall.R
\name{cuedRecall}
\alias{cuedRecall}
\title{PCR Cued Recall Simulation}
\usage{
cuedRecall(x, cue = 1, ...)
}
\arguments{
\item{x}{A PCR model object}

\item{...}{Additional arguments that may be passed to specific methods}
}
\description{
PCR Cued Recall Simulation
}

