% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/memories.R
\name{forget}
\alias{forget}
\alias{forget_beta}
\alias{forget_binomial}
\title{Forget}
\usage{
forget_binomial(mem, FR)

forget_beta(mem, FR)
}
\arguments{
\item{mem}{Matrix of memory strengths.}

\item{FR}{Forgetting Rate. Gives the probability of forgetting a feature}
}
\value{
Matrix of memory strengths updated after forgetting
}
\description{
Reduce the number of memory features activated by retrieval cues in the PCR
model.
}

