% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/recall.R
\name{freeRecall}
\alias{freeRecall}
\title{PCL Free Recall Simulation}
\usage{
freeRecall(mem, thresh, space = NULL, Tmin = NULL, Tmax = NULL,
  Time = NULL, lambda = NULL)
}
\arguments{
\item{mem}{}

\item{thresh}{}

\item{space}{}

\item{Tmin}{}

\item{Tmax}{}

\item{Time}{}

\item{lambda}{}
}
\value{
If time parameters are used, list of 3 matrices representing
 accuracy, reaction time, and serial order. If time parameters are not used, list of 2
 matrices representing accuracy and serial order.
}
\description{
PCL Free Recall Simulation
}

