% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/recall.R
\name{freeRecall}
\alias{freeRecall}
\title{PCL Free Recall Simulation}
\usage{
freeRecall(mem, thresh, space = NULL, Tmin = NULL, Tmax = NULL,
  Time = NULL, lambda = NULL)
}
\arguments{
\item{mem}{}

\item{thresh}{}

\item{space}{}

\item{Tmin}{}

\item{Tmax}{}

\item{Time}{}

\item{lambda}{}
}
\value{
If time parameters are used, list of 4 matrices representing
 accuracy, recoverability, reaction time, and serial order.
 If time parameters are not used, returns list of 2 matrices representing
 accuracy and serial order.
}
\description{
PCL Free Recall Simulation
}

