% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/memories.R
\name{study}
\alias{study}
\alias{study_beta}
\alias{study_binomial}
\title{Study}
\usage{
study(...)

study_beta(mem, nFeatures, LR, FR = NULL)

study_binomial(mem, nFeatures, LR, FR = NULL)
}
\arguments{
\item{...}{Arugments passed to approriate functions based on type-checking. See details.}

\item{mem}{Matrix of memory strengths.}

\item{nFeatures}{Number of features the memory for each item has}

\item{LR}{Learning Rate. Gives the probability of learning a new features
through practice}

\item{FR}{Forgetting Rate. Gives the probability of forgetting a feature}

\item{mem}{Matrix of memory strengths.}

\item{nFeatures}{Number of features the memory for each item has}

\item{LR}{Learning Rate. Gives the probability of learning a new features
through practice}

\item{FR}{Forgetting Rate. Gives the probability of forgetting a feature}
}
\value{
Matrix of memory strengths updated after restudying

Matrix of memory strengths updated after restudying
}
\description{
Effects of studying on un-encoded features
}
\details{
test will attempt to type-check the mem matrix and call the
appropriate function (study_beta for doubles,study_binomial for integers).
}
\section{Functions}{
\itemize{
\item \code{study_beta}: study_beta is for continous data

\item \code{study_binomial}: study_binomial is for discrete features
}}

