% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/memories.R
\name{study}
\alias{study}
\title{study}
\usage{
study(mem, nFeatures, LR, FR = NULL)
}
\arguments{
\item{mem}{Matrix of memory strengths.}

\item{nFeatures}{Number of features the memory for each item has}

\item{LR}{Learning Rate. Gives the probability of learning a new features
through practice}

\item{FR}{Forgetting Rate. Gives the probability of forgetting a feature}
}
\value{
Matrix of memory strengths updated after restudying
}
\description{
study
}

