% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCR.R
\name{updatePCRparams}
\alias{updatePCRparams}
\title{updatePCRparams}
\usage{
updatePCRparams(x, new)
}
\description{
Update PCRparams object and method with new parameter values
}
\examples{
param_list <- list(ER=.6, LR=.15, FR=.1, TR=.1, TV = .05, Tmax = 75, Tmin = 0.5, lambda = .5)

beta_timed <- initPCRparams(params = param_list, distribution = "beta", nItems = 20,
                           nSims = 1000, nFeatures = 100, time = 10)
new <- updatePCRparams(a, new = list(ER = .1, LR = .75))
}

